Split off ResolvedConfig from ImageCustomizerParameters. (#471)
Split the `ImageCustomizerParameters` into two, with a new
`ResolvedConfig` struct taking all the merged config values and any
generated values, and leaving `ImageCustomizerParameters` (renamed to
`imageMetadata`) to handle passing around image metadata. As part of
this, some redundant values in `ResolvedConfig` are being changed from
fields to functions.

This change is being done in preparation for the change that will merge
the config merging logic into the `ValidateConfig` function.

Also, handle the merging of the packages snapshot time when generating
the `ResolvedConfig` instead of deep within the code stack, to match how
other CLI parameters are handled.

Also, refactor some functions to take `ResolvedConfig` instead of a long
list of parameters, to help simplify code. Also, this will make the
upcoming Hierarchical Config changes a little eaiser.