Remove requirement for /boot partition for usr verity. (#496)
When using usr verity, the OS image is not required to have a separate
/boot partition. But Image Customizer's implementation has some flaws
preventing this at the moment.

This change is primarily focused on using the `/etc/fstab` file to find
the `/boot` directory instead of relying on the `grub.cfg` file on the
ESP partition. We already use the fstab file to discover the partition
layout. So, it is goodness to use it everywhere instead of having
multiple mechanisms. Also, this will make it easier to produce GRUB-less
OS images.

The primary fix comes from knowing if the /boot directory sits on its
own partition or if it is a subdirectory of the root partition. This
ensures that the code searches for files (e.g. `grub.cfg`) in the
correct locations.

This change switches the usr verity tests to use parition layouts
without a separate /boot partition. Whereas, the root verity tests keep
the separate /boot partition since it is required for root verity.