Reduce redundant partition discovery.
Currently, we have a `connectToExistingImage` function that tries to
discover the partition layout (by looking for the /etc/fstab file) and
then mounting all the partitions. This logic makes sense when we are
first connecting to a base image. But if we need to reconnect to the
image later in the customization process (e.g. to collect COSI
metadata), then we reuse the existing discovered partition layout
instead of trying to rediscover it.

This is mainly being done to allow users to mess with the fstab file
without it negatively affecting the customization process.