Fix CodeQL SM03116
The VMTests test suite uses SSH to connect to the QEMU VMs to verify
the OS image's contents. CodeQL is complaining that this code is
accepting unknown host keys. However, the VMs are being generated on
the fly. So, there is no way to pre-learn the SSH host keys. Also, this
is test code. Also, the SSH connection is occuring in a private virtual
network. Therefore, suppress the warning.